Remove `common` module
